pyproject.toml
[project]
name = "tlama-caller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "onesignal-python-api @ git+https://github.com/onesignal/onesignal-python-api.git",
    "customtkinter>=5.2.0",
    "Pillow>=10.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["."]
omit = ["tests/*", ".venv/*", "**/__pycache__/*"]
